Support recent versions of pip.
The old setup.py imported a requirement string parsing
util from `pip.req`. That package is no longer available in
pip > 9.0.3. That version of pip is now very old and outdated.

In newer versions of pip, that parsing util is not part
of the public API. Fortunately, setuptools itself provides
a standard, public method for parsing requirements strings.

This change switches the code to use that setuptools method.

As a result, hgvs can now be installed using modern versions
of pip.